- Moved the modal window for feedback on the event to the directory of modal windows
